Solved collapse problem with text annotations and changed arrow style for gene features.
Modified the collapse function to avoid calculating text bounding boxes for feature name label. Neither with matplotlib Text patch nor with the more advanced ax.annotate method we can retrieve the bounding box, there are some problems with the dpi of the figure that is not plotted yet. I just commented out the collision calculation for text annotations. Also I modified the drawing of feature names.

Modified the arrow style of the gene features.